e5422: add bin2 cpu support

+++ drivers/devfreq/exynos5422_ppmu.h
@@ -22,8 +22,8 @@ struct exynos5_ppmu_handle;
 struct exynos5_ppmu_handle *exynos5_ppmu_get(void);
 void exynos5_ppmu_put(struct exynos5_ppmu_handle *handle);
 int exynos5_ppmu_get_busy(struct exynos5_ppmu_handle *handle,
-	enum exynos_ppmu_sets filter, unsigned int *ccnt,
-	unsigned long *pmcnt);
+	enum exynos_ppmu_sets filter, unsigned long long *ccnt,
+	unsigned long long *pmcnt);
 void exynos5_ppmu_trace(void);
 
 #endif /* __DEVFREQ_EXYNOS5422_PPMU_H */
